requirements.txt
psutil==5.9.4
pycryptodomex==3.16.0
pywin32==305
requests==2.21.0
nuitka